CODE/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
python-dotenv==1.1.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.14.1
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
asgiref==3.9.1
diff-match-patch==20241021
Django==5.2.4
django-cors-headers==4.7.0
django-import-export==4.3.8
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.1
asgiref==3.9.1
diff-match-patch==20241021
Django==5.2.4
django-cors-headers==4.7.0
django-import-export==4.3.8
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
openpyxl==3.1.5
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.1
